Move utility consumers into separate module
